Move more code on Common to reduce duplication.
