+++ .idea/runConfigurations/Fabric_Client___fabric__fabric.xml
@@ -1,9 +1,9 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" factoryName="Application" name="Fabric Client (:fabric)" type="Application">
     <option name="MAIN_CLASS_NAME" value="net.fabricmc.devlaunchinjector.Main"/>
-    <module name="Dependent-1.20.1.fabric.main"/>
+    <module name="DangerClose-1.20-1.20.2.fabric.main"/>
     <option name="PROGRAM_PARAMETERS" value=""/>
-    <option name="VM_PARAMETERS" value="-Dfabric.dli.config=D:\JAVA\Mods-For-Lupin\_MonoLibDependent\Dependent-1.20.1\fabric\.gradle\loom-cache\launch.cfg -Dfabric.dli.env=client -Dfabric.dli.main=net.fabricmc.loader.impl.launch.knot.KnotClient"/>
+    <option name="VM_PARAMETERS" value="-Dfabric.dli.config=C:\Users\ricca\Desktop\DangerClose-1.20-1.20.2\fabric\.gradle\loom-cache\launch.cfg -Dfabric.dli.env=client -Dfabric.dli.main=net.fabricmc.loader.impl.launch.knot.KnotClient"/>
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/fabric/run/"/>
     <method v="2">
       <option enabled="true" name="Make"/>
